Add bionic vsnprintf and zygote whitelist shim
vsnprintf:
The graphics stack triggers a fortify assert in vsnprintf starting
with O. It tries to read past the end of a string. This shim
overrides vsnprintf and reestablishes the behaviour from N.

Zygote Whitelist:
The graphics stack leaves open file descriptors. Unless whitelisted
zygote will assert and crash when this happens. This shim overloads
the function that checks the whitelist and adds entries to it.

Change-Id: I837af11c6eb6afc948d6a6e8f320e05122e4128c